Update the bot

+++ src/commands/utility/say.js
@@ -2,8 +2,13 @@ module.exports = {
   name: 'say',
   permsWhitelist: ['ADMINISTRATOR'],
   execute(message, args, client) {
-    const channel = message.mentions.channels.first()
+    const channelID = args[0].replace(/[<@!&#>\\]/g, '')
     args.shift()
+
+    if(message.guild.id !== channel?.guild.id && !client.isOwner(message)) {
+      message.reply(client.getString('say.anotherServer', { locale: message }))
+      return
+    }
 
     if(message.deletable) message.delete()
     if(channel) channel.send(args.join(' '))
